unsubscribe()
Several improvements had to land to implement unsubscribing. HistoryCache was broken into ReaderHistoryCache and WriterHistoryCache that both handle keyed and non-keyed topics. GAP submessages are parsed and sent, and inlineQoS and serializedKey DATA fields can be serialized. The code is specific to unsubscribe() right now and will be generalized a bit when unregistering our participant during shutdown is implemented.